+++ src/.internal/charts/stock/drawing/LabelSeries.ts
@@ -89,7 +89,7 @@ export class LabelSeries extends PolylineSeries {
 		const text = dataContext.text;
 		const template = dataContext.settings;
 
-		container.set("forceInactive", true);
+		//container.set("forceInactive", true);
 
 		if (template) {
 			const label = container.children.push(EditableLabel.new(this._root, {