+++ src/.internal/charts/stock/toolbar/DrawingControl.ts
@@ -853,7 +853,9 @@ export class DrawingControl extends StockControl {
 			if (isInited) {
 				this.getPrivate("eraserControl")!.set("active", false);
 			}
-			this.get("stockChart").set("drawingSelectionEnabled", false)			
+			const stockChart = this.get("stockChart");
+			stockChart.set("drawingSelectionEnabled", false)			
+			stockChart.unselectDrawings();
 			return;
 		}